cleanup

+++ controller/cmd/service-mirror/probe_manager.go
@@ -181,7 +181,7 @@ func extractProbeSpec(svc *corev1.Service) (*multicluster.ProbeSpec, error) {
 
 	return &multicluster.ProbeSpec{
 		Path:   path,
-		Port:   uint32(probePort),
+		Port:   probePort,
 		Period: time.Duration(probePeriodSeconds) * time.Second,
 	}, nil
